Move FENDataCreator class to board fen module
