+++ src/board/fen.js
@@ -15,6 +15,11 @@ limitations under the License.
 */
 
 
+const { Piece, KingCastleRoad } = require('../pieces/main');
+const { Relation } = require('../relations');
+const { SquareName } = require('../square');
+
+
 /**
  * FEN string data parser class.
  *
@@ -47,6 +52,102 @@ class FENData {
 }
 
 
+/** FEN string data creator class. */
+class FENDataCreator {
+
+  #pieces = {
+    [Piece.WHITE]: {
+      [Piece.PAWN]: 'P',
+      [Piece.KNIGHT]: 'N',
+      [Piece.BISHOP]: 'B',
+      [Piece.ROOK]: 'R',
+      [Piece.QUEEN]: 'Q',
+      [Piece.KING]: 'K',
+    },
+    [Piece.BLACK]: {
+      [Piece.PAWN]: 'p',
+      [Piece.KNIGHT]: 'n',
+      [Piece.BISHOP]: 'b',
+      [Piece.ROOK]: 'r',
+      [Piece.QUEEN]: 'q',
+      [Piece.KING]: 'k',
+    },
+  };
+  #colors = {[Piece.WHITE]: 'w', [Piece.BLACK]: 'b'};
+  #castleRights = {
+    [Piece.WHITE]: {
+      [KingCastleRoad.SHORT]: 'K',
+      [KingCastleRoad.LONG]: 'Q',
+    },
+    [Piece.BLACK]: {
+      [KingCastleRoad.SHORT]: 'k',
+      [KingCastleRoad.LONG]: 'q',
+    },
+  };
+
+  /**
+   * Creation.
+   * @param {Board} board - Board instance.
+   */
+  constructor(board) {
+    this._board = board;
+    this.value = [
+      this._getPositionData(),
+      this.#colors[board.colors.current],
+      this._getCastleRightsData(),
+      board.enPassantSquare ? board.enPassantSquare.name.value : '-',
+      board.fiftyMovesRuleCounter.value.toString(),
+      board.movesCounter.value.toString(),
+    ].join(' ');
+  }
+
+  /**
+   * Get position data.
+   * @return {string} FEN position data.
+   */
+  _getPositionData() {
+    let data = [];
+    for (let number of SquareName.numbers) {
+      let rowData = [];
+      for (let symbol of SquareName.symbols) {
+        let square = this._board.squares[`${symbol}${number}`];
+        if (square.piece) {
+          rowData.push(this.#pieces[square.piece.color][square.piece.kind]);
+        } else {
+          rowData.push('0');
+        }
+      }
+      data.push(
+        rowData
+        .join('')
+        .replace(/0+/g, n => {return n.length})
+      );
+    }
+    return data.reverse().join('/');
+  }
+
+  /**
+   * Get castle rights data.
+   * @return {string} FEN castle rights data.
+   */
+  _getCastleRightsData() {
+    let data = [];
+    for (let color of Piece.ALL_COLORS) {
+      let king = this._board.kings[color];
+      if (king) {
+        for (let side of KingCastleRoad.ALL_SIDES) {
+          if (king.castle[side]) {
+            data.push(this.#castleRights[color][side]);
+          }
+        }
+      }
+    }
+    return data.join('') || '-';
+  }
+}
+
+
 module.exports = {
   FENData: FENData,
+  FENDataCreator: FENDataCreator,
 };
